Temporary repo

+++ Mhotivo.Interface/Interfaces/IAcademicYearDetailsRepository.cs
@@ -9,13 +9,13 @@ namespace Mhotivo.Interface.Interfaces
     public interface IAcademicYearDetailsRepository
     {
         AcademicYearDetail First(Expression<Func<AcademicYearDetail, AcademicYearDetail>> query);
-        AcademicYearDetail GetById(int id);
+        AcademicYearDetail GetById(long id);
         AcademicYearDetail Create(AcademicYearDetail academicYearToCreate);
         IQueryable<AcademicYearDetail> Query(Expression<Func<AcademicYearDetail, AcademicYearDetail>> expression);
         IQueryable<AcademicYearDetail> Filter(Expression<Func<AcademicYearDetail, bool>> expression);
-        IEnumerable<AcademicYearDetail> GetAllAcademicYearsDetails(int id);
+        IEnumerable<AcademicYearDetail> GetAllAcademicYearsDetails(long id);
         AcademicYearDetail Update(AcademicYearDetail itemToUpdate);
-        AcademicYearDetail Delete(int id);
+        AcademicYearDetail Delete(long id);
         void SaveChanges();
         IEnumerable<AcademicYearDetail> GetAllAcademicYearDetails();
         AcademicYearDetail FindByCourse(long id, long getMeisterId);
